Implement the layered rendering optimization
Rather than binding each face of the cubemap individually and then updating it, this optimization attaches the entire cubemap to an FBO, then renders into each of the six faces of the FBO by changing gl_Layer inside a geometry shader and emitting a rectangle to be rendered onto that layer (face) of the cubemap.

Associated per-face calculations are also performed in the geometry shader (per output vertex) and passed on to the fragment shader.

Implementing this optimization allows doubling the width and the height of the underlying cubemap texture without having performance dip substantially below 60FPS. This enables a much higher image density for the reaction diffusion simulation